// featureclient/types/websocket.ts
export interface ClientDebugSettings {
  enabled: boolean;
  enable_websocket_debug: boolean;
  enable_data_debug: boolean;
  log_binary_headers: boolean;
  log_full_json: boolean;
}